Traversy Views folder
This is the views folder which is part of the model, view, controller layout. It is taken from Traversy Media's website which can be explained on his laravel youtube series. Any HTML to be made should go in here. If you need to add something which is unique to each page and not going to be inherited by any other page, then put it in the 'pages' folder. Otherwise, put it in the 'inc' folder and then type @include(folder.filename) inside layouts.app(layout>app.blade.php).